combine suites
